requirements.txt
streamlit
langchain
langchain-openai
openai
python-dotenv
requests
